Remove project funding source and move the model to the funding app
