+++ funding/apps.py
@@ -1,0 +1,1 @@
+from django.apps import AppConfig
