+++ project/forms.py
@@ -37,7 +37,6 @@ class ProjectAdminForm(forms.ModelForm):
             'pi',
             'tech_lead',
             'category',
-            'funding_source',
             'start_date',
             'end_date',
             'economic_user',
@@ -113,7 +112,6 @@ class ProjectCreationForm(forms.ModelForm):
             'institution_reference',
             'department',
             'pi',
-            'funding_source',
             'start_date',
             'end_date',
             'requirements_software',
